requirements.txt
segno==1.4.1
Unidecode==1.3.2
google-api-python-client==2.33.0
streamlit==1.5.0
google-auth==2.6.0
requests==2.27.1